+++ build.zig
@@ -18,11 +18,12 @@ pub fn build(b: *std.Build) void {
     // set a preferred release mode, allowing the user to decide how to optimize.
     const optimize = b.standardOptimizeOption(.{});
 
-    // Add zcrypto dependency
-    const zcrypto = b.dependency("zcrypto", .{
+    // Get dependencies
+    const zcrypto_dep = b.dependency("zcrypto", .{
         .target = target,
         .optimize = optimize,
     });
+    const zcrypto_mod = zcrypto_dep.module("zcrypto");
 
     // This creates a module, which represents a collection of source files alongside
     // some compilation options, such as optimization mode and linked system libraries.
@@ -43,7 +44,7 @@ pub fn build(b: *std.Build) void {
         // which requires us to specify a target.
         .target = target,
         .imports = &.{
-            .{ .name = "zcrypto", .module = zcrypto.module("zcrypto") },
+            .{ .name = "zcrypto", .module = zcrypto_mod },
         },
     });
 
@@ -85,7 +86,7 @@ pub fn build(b: *std.Build) void {
                 // can be extremely useful in case of collisions (which can happen
                 // importing modules from different packages).
                 .{ .name = "zsig", .module = mod },
-                .{ .name = "zcrypto", .module = zcrypto.module("zcrypto") },
+                .{ .name = "zcrypto", .module = zcrypto_mod },
             },
         }),
     });